+++ ReleaseNotes/release_notes.md
@@ -4,6 +4,8 @@
   1. rewrite `statCovs` method to get statistics and covariances for trees and datastes  
   1. add `smattrix` method to `linalgt`-objects 
   1. improve printout of marices 
+  1. add `mahalanobis` distance 
+  1. improve `SVectorWithErrors`
 
 ## Backward incompatible: